HHH-12979 ConfigHelper is deprecated so let's not use it
ClassLoaderService provides what we need, except for the classpath://
URI resolution.

We don't support classpath:// elsewhere in the code, we do an attempt of
classpath resolution anyway.

So basically:
- if people uses a /simple/path/to/config/file, a classpath resolution
will be attempted;
- if people wants to use classpath:// (which is useless), they need
their framework to declare it (e.g. as in Spring). It's probably not a
good idea to try to override the framework behavior.